pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jao_scraper"
version = "1.1.0"
description = "Web scraping framework for JAO Publication Tool and other date-based data sources"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Henry Baker"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

dependencies = [
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "pandas>=2.1.0",
    "webdriver-manager>=4.0.0",
    "typer>=0.9.0",
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.2",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    # Code formatting and linting
    "black>=24.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "flake8-docstrings>=1.7.0",
    "flake8-bugbear>=24.1.0",
    "flake8-comprehensions>=3.14.0",
    # Type checking
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.0",
    # Security scanning
    "bandit[toml]>=1.7.6",
    "safety>=3.0.0",
    # Pre-commit hooks
    "pre-commit>=3.6.0",
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.urls]
Homepage = "https://github.com/henrybaker/jao_scraper"

# Black configuration - Python code formatter
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | __pycache__
  | \.eggs
  | \.pytest_cache
)/
'''

# isort configuration - Import sorting
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".venv", "venv", "build", "dist"]

# Mypy configuration - Static type checking
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_strict_optional = true
warn_unused_ignores = true
follow_imports = "silent"
show_column_numbers = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "example_usage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "standalone_jao_scraper"
ignore_errors = true

[[tool.mypy.overrides]]
module = "jao_api_scraper"
ignore_errors = true

# Bandit configuration - Security linting
[tool.bandit]
exclude_dirs = [
    "tests",
    ".venv",
    "venv",
]
skips = [
    "B101",  # assert_used - OK for development/testing
]

# Pytest configuration - Testing framework
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/example_usage.py",
    "*/standalone_jao_scraper.py",
    "*/jao_api_scraper.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
